Got User Auth Working.
I setup the user authentication. It will require a few things on the server side routes to get working when ever we set our routes up. I moved the routes file from the client side to the server side and added my test file there that can be used as a template. Also I added a user model but it was purely for test purposes and we can ignore it and build better models for the actual app. One more thing is I had to comment out the API.js in utils and index.js in routes because it was causing the server to crash. I think it is because none of these routes actually do anything yet.